services/email-service/requirements.txt
appnope==0.1.4
asttokens==2.4.1
Authlib==1.3.2
beautifulsoup4==4.10.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
falcon==4.0.2
falcon-cors==1.1.7
Flask==3.1.0
fonttools==4.55.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.0.2
kafka-python-ng==2.2.3
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pytz==2021.1
pyzmq==26.2.0
requests==2.26.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.17
wcwidth==0.2.13
Werkzeug==3.1.3
